oxigraph/oxigraph#1560 Move GeoSPARQL extension functions into spareval behind `geosparql` feature
