+++ lib/spargeo/src/vocab.rs
@@ -1,77 +1,11 @@
 //! GeoSPARQL 1.1 vocabulary constants.
 //!
-//! Re-exports every IRI this crate needs beyond the wkt and geoJSON literal
-//! datatypes already exposed by [`crate::geosparql`]. The goal is to let
-//! downstream consumers such as the optional bridge module reference core
+//! Re-exports every IRI that lives in [`spareval::geosparql::vocab`] and adds
+//! the `geosparql.ttl` ontology stub that only the bridge module in this
+//! crate needs. The goal is to let downstream consumers reference core
 //! GeoSPARQL predicates without hard coding the IRI strings.
 
-use oxrdf::NamedNodeRef;
-
-/// `geo:` namespace prefix used throughout the GeoSPARQL 1.1 ontology.
-pub const GEO_NS: &str = "http://www.opengis.net/ont/geosparql#";
-
-/// Class `geo:Feature`.
-pub const FEATURE: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#Feature");
-
-/// Class `geo:Geometry`.
-pub const GEOMETRY: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#Geometry");
-
-/// Class `geo:SpatialObject`.
-pub const SPATIAL_OBJECT: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#SpatialObject");
-
-/// Property `geo:hasGeometry`.
-pub const HAS_GEOMETRY: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#hasGeometry");
-
-/// Property `geo:hasDefaultGeometry`.
-pub const HAS_DEFAULT_GEOMETRY: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#hasDefaultGeometry");
-
-/// Property `geo:defaultGeometry`.
-pub const DEFAULT_GEOMETRY: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#defaultGeometry");
-
-/// Property `geo:asWKT`.
-pub const AS_WKT: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#asWKT");
-
-/// Property `geo:asGeoJSON`.
-pub const AS_GEO_JSON: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#asGeoJSON");
-
-/// Property `geo:hasSerialization`.
-pub const HAS_SERIALIZATION: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#hasSerialization");
-
-/// Datatype `geo:wktLiteral`.
-pub const WKT_LITERAL: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#wktLiteral");
-
-/// Datatype `geo:geoJSONLiteral`.
-pub const GEO_JSON_LITERAL: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#geoJSONLiteral");
-
-/// Simple Features topology predicates used by the bridge when materialising
-/// relations between feature geometries.
-pub const SF_WITHIN: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfWithin");
-pub const SF_CONTAINS: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfContains");
-pub const SF_EQUALS: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfEquals");
-pub const SF_OVERLAPS: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfOverlaps");
-pub const SF_TOUCHES: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfTouches");
-pub const SF_CROSSES: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfCrosses");
-pub const SF_DISJOINT: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfDisjoint");
-pub const SF_INTERSECTS: NamedNodeRef<'static> =
-    NamedNodeRef::new_unchecked("http://www.opengis.net/ont/geosparql#sfIntersects");
+pub use spareval::geosparql::vocab::*;
 
 /// GeoSPARQL 1.1 ontology stub shipped with the crate.
 ///
